requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
distlib==0.3.6
Django==4.1.3
filelock==3.8.0
idna==3.4
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==1.5.3
platformdirs==2.5.2
protobuf==3.20.3
pyfiglet==0.8.post1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.15.0
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.16.6
